refactor for event generator
