+++ src/app/core/event_samplers/gaussian_poisson.py
@@ -1,0 +1,7 @@
+"""
+event sampler in the case of gaussian distribution
+for concurrent user and poisson distribution for rqs per minute per user.
+The rationale behind this choice is about considering scenario
+with variance bigger or smaller w.r.t the one inherited from
+the Poisson distribution
+"""
