+++ src/app/core/event_samplers/poisson_poisson.py
@@ -1,0 +1,81 @@
+"""
+event sampler in the case of poisson distribution
+both for concurrent user and rqs per minute per user
+"""
+
+import math
+from collections.abc import Iterator
+
+import numpy as np
+
+from app.config.constants import TimeDefaults
+from app.core.event_samplers.common_helpers import (
+    poisson_variable_generator,
+    uniform_variable_generator,
+)
+from app.schemas.simulation_input import SimulationInput
+
+
+def poisson_poisson_sampling(
+    input_data: SimulationInput,
+    *,
+    simulation_time_second: int = TimeDefaults.SIMULATION_HORIZON.value,
+    sampling_window_s: int = TimeDefaults.SAMPLING_WINDOW.value,
+    rng: np.random.Generator | None = None,
+) -> Iterator[float]:
+    """
+    Yield inter-arrival gaps (seconds) for the compound Poisson-Poisson process.
+
+    Algorithm
+    ---------
+    1. Every *sampling_window_s* seconds, draw
+         U ~ Poisson(mean_concurrent_user).
+    2. Compute the aggregate rate
+         Λ = U * (mean_req_per_minute_per_user / 60)  [req/s].
+    3. While inside the current window, draw gaps
+         Δt ~ Exponential(Λ)   using inverse-CDF.
+    4. Stop once the virtual clock exceeds *simulation_time_second*.
+    """
+    rng = rng or np.random.default_rng()
+
+    # λ_u : mean concurrent users per window
+    mean_concurrent_user = float(input_data.avg_active_users.mean)
+
+    # λ_r / 60 : mean req/s per user
+    mean_req_per_sec_per_user = (
+        float(
+            input_data.avg_request_per_minute_per_user.mean)
+        / TimeDefaults.MIN_TO_SEC.value
+    )
+
+    now = 0.0                 # virtual clock (s)
+    window_end = 0.0          # end of the current user window
+    lam = 0.0                 # aggregate rate Λ (req/s)
+
+    while now < simulation_time_second:
+        # (Re)sample U at the start of each window
+        if now >= window_end:
+            window_end = now + float(sampling_window_s)
+            users = poisson_variable_generator(mean_concurrent_user, rng)
+            lam = users * mean_req_per_sec_per_user
+
+        # No users → fast-forward to next window
+        if lam <= 0.0:
+            now = window_end
+            continue
+
+        # Exponential gap from a protected uniform value
+        u_raw = max(uniform_variable_generator(rng), 1e-15)
+        delta_t = -math.log(1.0 - u_raw) / lam
+
+        # End simulation if the next event exceeds the horizon
+        if now + delta_t > simulation_time_second:
+            break
+
+        # If the gap crosses the window boundary, jump to it
+        if now + delta_t >= window_end:
+            now = window_end
+            continue
+
+        now += delta_t
+        yield delta_t
