+++ src/app/schemas/simulation_input.py
@@ -19,7 +19,7 @@ class RVConfig(BaseModel):
         ) -> float:
         """Ensure `mean` is numeric, then coerce to float."""
         err_msg = "mean must be a number (int or float)"
-        if not isinstance(v, float):
+        if not isinstance(v, (float, int)):
             raise ValueError(err_msg)  # noqa: TRY004
         return float(v)
